moved ProjectItem to own file
